feat: Created AdaptiveContainer
